# openapi.yaml
openapi: 3.0.0
info:
  title: BRDG JPMP API
  description: API for the BRDG JPMP kiosk application
  version: 1.0.0
servers:
  - url: https://api.jpmp.brdg.jarv.us
    description: Production backend
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /items:
    get:
      summary: Get all items
      description: Returns a list of all items
      responses:
        "200":
          description: A list of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
    post:
      summary: Create a new item
      description: Creates a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kiosk_type
                - item_type
                - name
                - slug
              properties:
                kiosk_type:
                  type: string
                  enum: [sweet, juice]
                item_type:
                  type: string
                  enum: [sweet, juice, gift]
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                available:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /items/kiosk/{type}:
    get:
      summary: Get items by kiosk type
      description: Returns a list of items for a specific kiosk type
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [sweet, juice]
      responses:
        "200":
          description: A list of items for the specified kiosk type
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Item"
        "400":
          description: Invalid kiosk type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /items/{slug}:
    get:
      summary: Get a specific item by slug or ID
      description: Returns a specific item by slug or numeric ID (for backwards compatibility)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
          description: Item slug (string) or ID (integer)
      responses:
        "200":
          description: The requested item
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: "#/components/schemas/Item"
        "400":
          description: Invalid item slug or ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update an item by slug or ID
      description: Updates an existing item by slug or numeric ID (for backwards compatibility)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
          description: Item slug (string) or ID (integer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                  description: New slug for the item
                description:
                  type: string
                available:
                  type: boolean
                item_type:
                  type: string
                  enum: [sweet, juice, gift]
      responses:
        "200":
          description: Item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete an item by slug or ID
      description: Deletes an existing item by slug or numeric ID (for backwards compatibility)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
          description: Item slug (string) or ID (integer)
      responses:
        "200":
          description: Item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Invalid item slug or ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
          description: Unique identifier for the item
        kiosk_type:
          type: string
          enum: [sweet, juice]
        item_type:
          type: string
          enum: [sweet, juice, gift]
        name:
          type: string
        description:
          type: string
        available:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            status:
              type: integer